+++ src/entities/Album.java
@@ -1,4 +1,4 @@
-package api.src.entities;
+package src.entities;
 
 import se.michaelthelin.spotify.model_objects.specification.ArtistSimplified;
 import se.michaelthelin.spotify.model_objects.specification.Image;
