+++ src/entities/factories/AlbumFactory.java
@@ -1,4 +1,6 @@
-package api.src.entities;
+package src.entities.factories;
+
+import src.entities.Album;
 
 public interface AlbumFactory {
     Album create(String id);
