+++ src/entities/factories/SongFactory.java
@@ -1,4 +1,6 @@
-package api.src.entities;
+package src.entities.factories;
+
+import src.entities.Song;
 
 public interface SongFactory {
     Song create(String id);
