+++ src/interface_adapter/take_quiz/TakeQuizController.java
@@ -1,4 +1,4 @@
-package api.src.interface_adapter.take_quiz;
+package src.interface_adapter.take_quiz;
 
 public class TakeQuizController {
 }
