+++ src/use_case/take_quiz/TakeQuizInputBoundary.java
@@ -1,4 +1,4 @@
-package api.src.use_case.take_quiz;
+package src.use_case.take_quiz;
 
 public class TakeQuizInputBoundary {
 }
