+++ src/view/TakeQuizView.java
@@ -1,4 +1,4 @@
-package api.src.view;
+package src.view;
 
 public class TakeQuizView {
 }
